totalne zmiany w serwerze streamlinka

+++ SL/configurator/scripts/getOnlineFiles.sh
@@ -11,5 +11,7 @@ unzip -q ~/streamlink-master.zip
 #cp -rf ~/streamlink-master/src/streamlink/* $myAbsPath/../../streamlink/
 #cp -rf ~/streamlink-master/src/streamlink_cli/* $myAbsPath/../../streamlink_cli/
 cp -rf ~/streamlink-master/src/streamlink_cli/* $myAbsPath/../plugins/streamlink_cli/
-
+if [ `grep -c '#j00zek_patch 2' < $myAbsPath/../plugins/streamlink_cli/main.py` -eq 0 ];then
+ sed -i 's/\(.*\)\(log\.info."Stream ended".*\)/\1\2\n\1if args.player_external_http: sys.exit() #j00zek_patch 2/' $myAbsPath/../plugins/streamlink_cli/main.py
+fi
 rm -rf ~/streamlink-master*
